Safari-611.1.10/LayoutTests/platform/ios/editing/pasteboard/copy-paste-across-shadow-boundaries-3-expected.txt

This tests copying and pasting content across shadow boundaries.
To test manually, copy text blow starting from "WebKit" ending with "rocks", and paste into the green box below. All the text shoul be copied & pasted.

pasted html:
| <span>
|   id="host"
|   style="-webkit-text-size-adjust: auto;"
|   <span>
|     style="display: contents;"
|     "world WebKit"
| <span>
|   style="-webkit-text-size-adjust: auto;"
|   " rocks<#selection-caret>"

text/plain:
| "world WebKit rocks"
